+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Script to delete current duplicate authors/contributors in the PSQL database
 - Web scraper script for the Chicago ISAC catalog of publications
 - Mapping and process for Chicago ISAC catalog of publications
+- Unit tests for Chicago ISAC mapping and process
 ### Fixed
 - Modify agentParser method to reduce number of future duplicate agents
 - Install `wheel` with pip to fix fasttext build